Update nPose Core.lsl
Change in LINKMSG processing to use menu toucher key if none are included in the message rather than grabbing the last slot sitter key.

This is a fix for EXTERNAL_UTIL_REQUEST in the menu which became broken when toucherid was changed in the menu.  This also makes far more sense than grabbing the last slots list stride key.