Update nPose Slave.lsl
This takes care of issue #33 and lowers the timer to 0.1 seconds.  Removed the RezAdjuster function and replace it with RezAllAdjusters function and cleaned up the rest of the code.

While testing this new function I found that when changing poses, the Adjusters were always one step behind (had to select a new pose twice to get adjusters to move).  Found the issue in the Core script where it sent the command to re-rez adjusters before the slots list was updated.  Made the change in the Core to support this function.  Now all is well and the timer in the slave is set to 0.1 seconds.

Glad you didn't let me push this under the rug and left it as an issue to check out.